fix(DF-74): 위치 계산 유틸 함수 수정 - 산책로를 생성할 때 소수점 2번째자리까지 반올림 한  distance값을 반환하도록 수정
