+++ src/pages/newway/index.tsx
@@ -1,58 +1,18 @@
-import React, { useState, useRef, useEffect } from "react";
+import { useEffect, useRef, useState } from "react";
 import styled from "styled-components";
+import { calculateDistance } from "src/utils/calculateDistance";
+import TrackingMapTest from "../../components/map/TrackingMap";
+import SmallButton from "src/components/button/SmallButton";
+import ConfirmationModal from "src/components/modal/ConfirmationModal";
+import useWatchLocation from "src/hooks/useWatchLocation";
+import TrailInfo from "src/components/newway_register/TrailInfo";
 import { useNavigate } from "react-router-dom";
-import { TrackingMap } from "../../components/map/TrackingMap";
-import TrailInfo from "../../components/newway_register/TrailInfo";
-import SmallButton from "../../components/button/SmallButton";
-import ConfirmationModal from "../../components/modal/ConfirmationModal";
+import { BiCurrentLocation } from "react-icons/bi";
 import AppBar from "src/components/appBar";
-
-const PageContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  height: calc(100dvh - 56px);
-  margin-top: 56px;
-`;
-
-const MapContainer = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 0;
-`;
-
-const OverlayContainer = styled.div`
-  position: absolute;
-  top: -56px;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 10;
-  pointer-events: none;
-`;
-
-const TopOverlay = styled.div`
-  background-color: rgba(255, 255, 255, 0.8);
-  margin: 1rem;
-  border-radius: 0.5rem;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const BottomOverlay = styled.div`
-  position: absolute;
-  bottom: calc(0% + 80px);
-  right: calc(0% + 30px);
-  background-color: white;
-  box-shadow: 0 3px 5px rgba(0, 0, 0, 0.5);
-  pointer-events: auto;
-  border-radius: 10px;
-`;
+import { drawPath } from "src/utils/drawPathUtils";
+import { uploadCourseImage } from "src/apis/walkway";
+import { useToast } from "src/hooks/useToast";
+import WaveTextLoader from "src/components/loading/WaveTextLoader";
 
 interface Location {
   lat: number;
@@ -65,141 +25,261 @@ interface PathData {
   duration: number;
   startTime: Date;
   endTime: Date;
+  pathImage: string;
+  courseImageId: number;
 }
 
-function NewWay() {
+// 산책중단 버튼 클릭시 팜업 모달 | 백버튼 클릭시 팝업하는 모달
+type ModalType = "stop" | "back" | null;
+
+const Container = styled.div`
+  position: relative;
+  width: 100%;
+  height: 100dvh;
+  max-width: 430px;
+  margin: 0 auto;
+`;
+
+const InfoContainer = styled.div`
+  position: absolute;
+  top: 20px;
+  left: 20px;
+  right: 20px;
+  background-color: rgba(255, 255, 255, 0.9);
+  border-radius: 10px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  z-index: 20;
+  display: flex;
+  justify-content: center;
+`;
+
+const ButtonContainer = styled.div`
+  position: absolute;
+  bottom: 100px;
+  right: 30px;
+  z-index: 1;
+`;
+
+const LocationButton = styled.button`
+  position: absolute;
+  bottom: 180px;
+  right: 30px;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  background-color: white;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: none;
+  cursor: pointer;
+  z-index: 1;
+
+  &:active {
+    background-color: #f0f0f0;
+  }
+`;
+
+export default function NewWay() {
   const navigate = useNavigate();
-  const pathCoordsRef = useRef<Location[]>([]);
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
+  const { showToast } = useToast();
   const [isWalking, setIsWalking] = useState(false);
-  const [distance, setDistance] = useState(0);
-  const [totalSeconds, setTotalSeconds] = useState(0);
+  const [modalType, setModalType] = useState<ModalType>(null);
+  const [userLocation, setUserLocation] = useState<Location | null>(null);
+  const [movingPath, setMovingPath] = useState<Location[]>([]);
+  const [distances, setDistances] = useState(0);
+  const [elapsedTime, setElapsedTime] = useState(0);
+  const timerRef = useRef<NodeJS.Timeout | null>(null);
+  const lastLocationRef = useRef<Location | null>(null);
+  const startTimeRef = useRef<Date | null>(null);
 
-  // 타이머 관련 ref
-  const startTimeRef = useRef<number | null>(null);
-  const timerIdRef = useRef<NodeJS.Timeout | null>(null);
-
-  // 타이머 업데이트 함수
-  const updateTimer = () => {
-    if (!startTimeRef.current) return;
-
-    const now = Date.now();
-    const diff = now - startTimeRef.current;
-    const seconds = Math.floor(diff / 1000);
-    setTotalSeconds(seconds);
+  const geolocationOptions = {
+    enableHighAccuracy: true,
+    maximumAge: 3000,
+    timeout: 3000,
   };
 
-  // 산책 시작/중단 핸들러
-  const handleButtonClick = () => {
-    if (!isWalking) {
-      // 산책 시작
-      setIsWalking(true);
-      startTimeRef.current = Date.now();
-      pathCoordsRef.current = [];
-      setDistance(0);
-      setTotalSeconds(0);
+  const { location, getLocation } = useWatchLocation(geolocationOptions);
 
-      // 타이머 시작
-      timerIdRef.current = setInterval(updateTimer, 1000);
-    } else {
-      // 산책 중단
-      setIsModalOpen(true);
-    }
-  };
+  useEffect(() => {
+    if (!location) return;
 
-  // 위치 업데이트
-  const handleLocationUpdate = (newLocation: Location) => {
-    pathCoordsRef.current.push(newLocation);
-  };
-
-  // 거리 업데이트
-  const handleDistanceUpdate = (newDistance: number) => {
-    setDistance((prev) => prev + newDistance);
-  };
-
-  // 산책 중단 확인
-  const handleStopWalk = () => {
-    // 타이머 정지
-    if (timerIdRef.current) {
-      clearInterval(timerIdRef.current);
-    }
-
-    const pathData: PathData = {
-      coordinates: pathCoordsRef.current,
-      totalDistance: Number(distance.toFixed(2)),
-      duration: totalSeconds,
-      startTime: new Date(startTimeRef.current!),
-      endTime: new Date(),
+    const newLocation = {
+      lat: location.latitude,
+      lng: location.longitude,
     };
 
-    console.log("산책 중단 시 저장된 경로 좌표:", {
-      좌표배열: pathData.coordinates,
-      총거리: pathData.totalDistance,
-      소요시간: pathData.duration,
-      시작시간: pathData.startTime,
-      종료시간: pathData.endTime,
-    });
+    setUserLocation(newLocation);
+    lastLocationRef.current = newLocation;
+  }, [location]);
 
-    setIsModalOpen(false);
-    setIsWalking(false);
+  useEffect(() => {
+    let interval: NodeJS.Timeout | null = null;
 
-    // 등록 페이지로 이동하면서 데이터 전달
-    navigate("/newway/registration", { state: pathData });
+    if (isWalking && lastLocationRef.current) {
+      interval = setInterval(() => {
+        if (lastLocationRef.current) {
+          setMovingPath((prev) => {
+            const newPath = [...prev, lastLocationRef.current!];
+            if (prev.length > 0) {
+              const newDistance = calculateDistance(newPath);
+              setDistances((prev) => prev + newDistance);
+            }
+            return newPath;
+          });
+        }
+      }, 3000);
+    }
+
+    return () => {
+      if (interval) {
+        clearInterval(interval);
+      }
+    };
+  }, [isWalking]);
+
+  const handleStartWalking = () => {
+    setIsWalking(true);
+    setMovingPath([]);
+    setDistances(0);
+    setElapsedTime(0);
+    startTimeRef.current = new Date();
+
+    timerRef.current = setInterval(() => {
+      setElapsedTime((prev) => prev + 1);
+    }, 1000);
   };
 
-  // 산책 계속하기
-  const handleContinueWalk = () => {
-    setIsModalOpen(false);
+  const handleStopRequest = () => {
+    if (elapsedTime < 60) {
+      // 산책 경과 시간이 1분 미만인 경우
+      showToast("1분 이상 산책해주세요.", "error");
+      return;
+    }
+    setModalType("stop");
+  };
+  const handleBackClick = () => {
+    setModalType("back");
   };
 
-  // 컴포넌트 언마운트 시 타이머 정리
+  const handleModalClose = () => {
+    setModalType(null);
+  };
+
+  const handleModalConfirm = async () => {
+    if (modalType === "stop") {
+      setIsWalking(false);
+
+      if (timerRef.current) {
+        clearInterval(timerRef.current);
+        timerRef.current = null;
+      }
+
+      try {
+        const pathImage = await drawPath(movingPath);
+
+        const blob = await fetch(pathImage).then((res) => res.blob());
+        const courseImageFile = new File([blob], "course-image.png", {
+          type: "image/png",
+        });
+
+        const courseImageId = await uploadCourseImage(courseImageFile);
+
+        const pathData: PathData = {
+          coordinates: movingPath,
+          totalDistance: distances,
+          duration: elapsedTime,
+          startTime:
+            startTimeRef.current || new Date(Date.now() - elapsedTime * 1000),
+          endTime: new Date(),
+          pathImage: pathImage,
+          courseImageId: courseImageId,
+        };
+
+        navigate("/newway/registration", {
+          state: { ...pathData, isEditMode: false },
+        });
+      } catch (error) {
+        console.error("이미지 업로드 실패:", error);
+      }
+    } else if (modalType === "back") {
+      navigate("/main");
+    }
+    setModalType(null);
+  };
+
+  const getModalType = (type: ModalType) => {
+    switch (type) {
+      case "stop":
+        return {
+          message: "산책을 중단하시겠습니까?",
+          cancelText: "계속하기",
+          confirmText: "중단하기",
+        };
+      case "back":
+        return {
+          message: `홈으로 돌아가시겠습니까?
+                    산책정보는 저장되지 않습니다.`,
+          cancelText: "취소",
+          confirmText: "확인",
+        };
+      default:
+        return {
+          message: "",
+          cancelText: "취소",
+          confirmText: "확인",
+        };
+    }
+  };
+
+  const handleUpdateLocation = () => {
+    getLocation();
+  };
+
   useEffect(() => {
     return () => {
-      if (timerIdRef.current) {
-        clearInterval(timerIdRef.current);
+      if (timerRef.current) {
+        clearInterval(timerRef.current);
       }
     };
   }, []);
 
+  if (!userLocation) {
+    return <WaveTextLoader>위치 정보를 불러오는 중...</WaveTextLoader>;
+  }
   return (
     <>
-      <AppBar onBack={() => navigate(-1)} title="산책로 등록" />
-      <PageContainer>
-        <MapContainer>
-          <TrackingMap
-            isTracking={isWalking}
-            onLocationUpdate={handleLocationUpdate}
-            onDistanceUpdate={handleDistanceUpdate}
+      <AppBar onBack={handleBackClick} title="산책로 등록" />
+      <Container>
+        <InfoContainer>
+          <TrailInfo duration={elapsedTime} distance={distances / 1000} />
+        </InfoContainer>
+
+        <TrackingMapTest userLocation={userLocation} movingPath={movingPath} />
+
+        {!isWalking && (
+          <LocationButton onClick={handleUpdateLocation}>
+            <BiCurrentLocation size={24} />
+          </LocationButton>
+        )}
+
+        <ButtonContainer>
+          <SmallButton
+            primaryText="산책 시작"
+            secondaryText="산책 중단"
+            isWalking={isWalking}
+            onClick={isWalking ? handleStopRequest : handleStartWalking}
           />
-        </MapContainer>
-
-        <OverlayContainer>
-          <TopOverlay>
-            <TrailInfo duration={totalSeconds} distance={distance} />
-          </TopOverlay>
-
-          <BottomOverlay>
-            <SmallButton
-              primaryText="산책 시작"
-              secondaryText="산책 중단"
-              isWalking={isWalking}
-              onClick={handleButtonClick}
-            />
-          </BottomOverlay>
-        </OverlayContainer>
+        </ButtonContainer>
 
         <ConfirmationModal
-          isOpen={isModalOpen}
-          onClose={handleContinueWalk}
-          onConfirm={handleStopWalk}
-          message="산책을 중단하시겠습니까?"
-          cancelText="계속하기"
-          confirmText="중단하기"
+          isOpen={modalType !== null}
+          onClose={handleModalClose}
+          onConfirm={handleModalConfirm}
+          {...getModalType(modalType)}
         />
-      </PageContainer>
+      </Container>
     </>
   );
 }
-
-export default NewWay;
